use proper type for form element.
coming up against https://bugs.php.net/bug.php?id=80490.
the short of it is that 0 is being treated as PARAM_RAW in PHP7.4 in a
switch/case statement within the clean_param() function.
This behaviour has changed in PHP 8+, leading to an exception being
raised since 0 is now rejected as invalid type value.
Changing the type to PARAM_RAW prevents this exception and is
PHP-version backwards compatible.